.history/requirements_20250718110714.txt
Flask
cohere
python-dotenv
Flask
Flask-SQLAlchemy
Flask-WTF
pandas
openpyxl
